+++ src/homework/Task1.java
@@ -1,0 +1,10 @@
+import java.time.LocalDateTime;
+
+public class Task1 {
+    public static void main(String[] args) {
+        String developerLastName = "Борук";
+        System.out.println(developerLastName);
+        LocalDateTime currentDateTime = LocalDateTime.now();
+        System.out.print(currentDateTime);
+    }
+}
